test: exclude untestable native provider logic from coverage reports
